requirements.txt
python-dotenv==1.0.1

langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.3
langchain-groq==0.1.10

chromadb==0.4.24
pypdf==4.2.0
sentence-transformers==2.7.0
huggingface-hub==0.23.4
torch>=2.0.0
